+++ common/__init__.py
@@ -3,9 +3,10 @@
 from .angle import THETA_DEGREES, circular_distance_deg, wrap_theta_deg
 from .data_types import (
     CandidateDirection, DecisionWindow, EnhancedAudio, IngestedAudioBlock,
-    PipelineStatus, SpatialResponse,
+    PipelineStatus, SpatialResponse, TrackedDirection,
 )
 from .geometry import MIC_POSITIONS_M, MicGeometry, physical_6plus1_geometry
+from .window_key import WindowKey
 
 __all__ = [
     "CandidateDirection",
@@ -17,6 +18,8 @@ __all__ = [
     "PipelineStatus",
     "SpatialResponse",
     "THETA_DEGREES",
+    "TrackedDirection",
+    "WindowKey",
     "circular_distance_deg",
     "physical_6plus1_geometry",
     "wrap_theta_deg",